Created an Settings subcomponent
