+++ src/features/nav/Nav.js
@@ -1,6 +1,5 @@
 import { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import SettingsIcon from '@mui/icons-material/Settings';
+import { useSelector } from 'react-redux';
 import MenuIcon from '@mui/icons-material/Menu';
 import CloseIcon from '@mui/icons-material/Close';
 import { Button, ButtonGroup, useMediaQuery } from '@mui/material';
@@ -10,14 +9,11 @@ import Play from 'features/nav/Play';
 import OpeningSearch from 'features/nav/OpeningSearch';
 import Database from 'features/nav/Database';
 import Training from 'features/nav/Training';
-import * as navConst from 'features/nav/navConst';
-import * as nav from 'features/nav/navSlice';
 import Inbox from './Inbox';
+import Settings from './Settings';
 
 const Nav = () => {
   const state = useSelector(state => state);
-
-  const dispatch = useDispatch();
 
   const maxWidth = {
     '900': useMediaQuery("(max-width:900px)"),
@@ -30,7 +26,7 @@ const Nav = () => {
     setHamburgerMenu(!hamburgerMenuOpen);
   }
 
-  
+
 
   const disabled = state.playMode.active &&
     state.playMode.accepted &&
@@ -44,7 +40,7 @@ const Nav = () => {
   return (
     <>
       <Button sx={{ display: `${maxWidth['900'] ? "block" : "none"}` }}
-        onClick={handleHamburgerClick} > {hamburgerMenuOpen ? <CloseIcon sx={{fontSize: 40}} /> : <MenuIcon sx={{fontSize: 40}} />}
+        onClick={handleHamburgerClick} > {hamburgerMenuOpen ? <CloseIcon sx={{ fontSize: 40 }} /> : <MenuIcon sx={{ fontSize: 40 }} />}
       </Button>
       <ButtonGroup
         orientation={maxWidth['900'] ? "vertical" : "horizontal"}
@@ -64,16 +60,8 @@ const Nav = () => {
         <OpeningSearch />
         <Database />
         <Training />
-        <Inbox/>
-        <Button
-          id="Nav-settings"
-          sx={{ borderRadius: 0 }}
-          variant={state.nav.name === navConst.SETTINGS ? "contained" : "text"}
-          startIcon={<SettingsIcon />}
-          onClick={() => dispatch(nav.settingsDialog({ open: true }))}
-        >
-          Settings
-        </Button>
+        <Inbox />
+        <Settings />
       </ButtonGroup>
     </>
   );
