+++ src/features/nav/Settings.js
@@ -1,0 +1,28 @@
+import React from 'react'
+import SettingsIcon from '@mui/icons-material/Settings';
+import { useDispatch, useSelector } from 'react-redux';
+import { Button } from '@mui/material';
+import * as navConst from 'features/nav/navConst';
+import * as nav from 'features/nav/navSlice';
+const Settings = () => {
+    const state = useSelector(state => state);
+
+    const dispatch = useDispatch();
+
+
+    return (
+        <>
+            <Button
+                id="Nav-settings"
+                sx={{ borderRadius: 0 }}
+                variant={state.nav.name === navConst.SETTINGS ? "contained" : "text"}
+                startIcon={<SettingsIcon />}
+                onClick={() => dispatch(nav.settingsDialog({ open: true }))}
+            >
+                Settings
+            </Button>
+        </>
+    )
+}
+
+export default Settings
